Use -xMMD instead of -xMD on OpenSolaris.
Using -xMD generates dependencies for all files including the
system files. This could point to some system .cc files and
was causing GNU make to try to build these files if the
dependency files existed. Using -xMMD generates only the
dependencies outside the system directories.

It's possible that the OpenSolaris make knows how to deal with
this correctly, but Icarus needs GNU make.